Separate struct __pos_operations to struct __pos_operations (in the include/oneapi/dpl/pstl/utils.h) and to struct __pos_operations_sycl (in the include/oneapi/dpl/pstl/hetero/utils_hetero.h) for sycl-specific operations
